runPseudotimeDE works with counts matrix
Added (initially failing) test for running on count matrix. Removed the subsetting `mat[x,]` in runPseudotimeDE() as we would need `mat[x,,drop=FALSE]` to avoid breaking pseudotimeDE(). Adding drop=FALSE might interfere with Seurat and SCE objects, thus rather not using it here. It's unclear to me why runPseudotimeDE was subsetting, as the tests on SCE and Seurat objects are not affected.